crash and chat fixes, chat and camera improvements
Crash fix. When the client was launched and no connection to the server
was established the client would crash. Now the client will print to the
console a message stating that no connection was made and the client is
now in stand alone mode.
Pressing the [t]key now properly sets focus to the chat input element.
Improvements to the location of the chat window.
Camera code has been much revamped adding a toggle between 3 POV setting
by pressing the [c]key